Add cudnn.BatchNormalization and cudnn.VolumetricBatchNormalization
