+++ README.md
@@ -1,21 +1,21 @@
 cudnn.torch
 ===========
 
-Torch7 FFI bindings for NVidia CuDNN (R4) kernels!
+Torch7 FFI bindings for NVIDIA cuDNN (R4) kernels!
 
 Modules are API compatible their [`nn`](https://github.com/torch/nn) equivalents. Fully unit-tested against `nn` implementations.
 Conversion between `nn` and `cudnn` is available through `cudnn.convert` function.
 
 #### Installation
 
-* Install CuDNN (version R4 EA)
-* Have at least Cuda 7.0
+* Install cuDNN (version R4 EA)
+* Have at least CUDA 7.0
 * Have `libcudnn.so` in your library path (Install it from https://developer.nvidia.com/cuDNN )
 
 #### Modules
 
 ```lua
--- All inputs have to be 3D or 4D(batch-mode), except ReLU, Tanh and Sigmoid
+-- All inputs have to be 3D or 4D(batch-mode), except ReLU, Tanh, Sigmoid, and BatchNormalization
 cudnn.SpatialConvolution(nInputPlane, nOutputPlane, kW, kH, [dW = 1], [dH = 1], [padW = 0], [padH = 0], [groups = 1])
 cudnn.SpatialMaxPooling(kW, kH, dW, dH, padW, padH)
 cudnn.SpatialAveragePooling(kW, kH, dW, dH, padW, padH)
@@ -91,13 +91,3 @@ For version CuDNN R3, checkout the branch **R3**
 R4 Release Notes:
 - Rather than resolving v3-v4 diffs, I have imported new cudnn.h with its entirety and converted comments and defines. This should be less error-prone.
 - addTensor_v2 uses changed to new AddTensor API.
-
-R4 TODO:
-per-activation BN code needs to be added (new .lua similar to SpatialBN.lua, as per Andrei:
-I believe we have at least one thing missing - per-activation BN (Torch implementation in nn.BatchNormalization.lua).
-What I believe we have now is an integration of implementation for nn.SpatialBatchNormalization.lua
-
-This is very similar to SpatialBatchNormalizaiton.lua but should use a different cudnnBatchNormalizationMode_t and tensor dimensions need to be adjusted accordingly.
-For Spatial BN normalization is performed over N with 1CHW result and for per-activation it's done over NHW with 1C11 result.
-
-Per-activation BN is only used after non-convolutional layers where spatially-invariant behavior is not expected.
