+++ VolumetricBatchNormalization.lua
@@ -1,0 +1,5 @@
+local VolumetricBatchNormalization =
+   torch.class('cudnn.VolumetricBatchNormalization', 'cudnn.BatchNormalization')
+
+VolumetricBatchNormalization.mode = 'CUDNN_BATCHNORM_SPATIAL'
+VolumetricBatchNormalization.nDim = 5
